requirements.txt
numpy>=2.1.0
Pillow>=10.4.0
opencv-python>=4.10.0.84
torch>=2.6.0
torchvision>=0.20.0
reedsolo==1.7.0
scikit-image>=0.24.0
click>=8.1.7
tqdm>=4.66.4
requests>=2.32.3
